scn-v2-demo/scn/requirements.txt
backports.shutil-get-terminal-size==1.0.0
boto3==1.9.75
botocore==1.12.75
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
docutils==0.14
ecdsa==0.13
enum34==1.1.6
envs==1.3
future==0.17.1
idna==2.6
ipython==5.7.0
ipython-genutils==0.2.0
jmespath==0.9.3
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycryptodome==3.3.1
Pygments==2.2.0
python-dateutil==2.7.5
python-jose-cryptodome==1.3.2
requests==2.18.4
s3transfer==0.1.13
scandir
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
warrant==0.6.1
wcwidth==0.1.7
python-dotenv
behave==1.2.6
selenium==3.141.0
webdriver-manager==3.4.2
pyYaml
